# docs/中医电子病历质控与标准化系统-openapi.yaml
openapi: 3.0.3
info:
  title: 中医电子病历质控与标准化系统
  description: |
    中医电子病历质控与标准化系统接口文档。

    ## 权限约定
    接口分四档权限，各接口在 description 首部以【权限：…】标注：
    - **仅管理员**：管理类操作（词典导入/回滚、数据清洗、数据集导出、日志审计、质控评分与重算）。
    - **登录即可**：只读查询与业务计算（术语查询/归一、统计看板、NLP 抽取、原始病历查看）。
    - **管理员 / 审核员**：复核相关接口（复核任务列表、提交复核）。
    - **公开**：用户登录、用户注册。
    审核员另受**数据域（行级权限）**约束：仅可读取 grade='待复核' 的病历；管理员为全库。
    执行顺序固定为「先数据域、后用户筛选」取交集，用户筛选条件不得绕过数据域。
    权限不足统一返回 HTTP 403（code=403）。
    ## 实体分层命名约定
    - **DTO**（后缀 DTO）：请求体入参模型，接口入参统一使用。
    - **VO**（后缀 VO）：响应体出参模型，接口出参统一使用。
    - **PO**（后缀 PO）：数据库持久层实体，内部使用，不暴露于接口，故本文档不列出。
    - **共享模型**（无后缀）：跨请求/响应复用的数据模型（如 StructuredData、Entity、Herb、RecordMetadata），见数据模型定义。
    ## 接口通用约定
    除标注「文件流下载」的接口外，所有接口统一返回 Result 包装 {"code":200,"msg":"string","data":any}。
    - code: 200成功；400参数非法；401未登录/token失效；403权限不足；404资源不存在；500服务异常。
    - 登录态：非登录接口需在请求头携带 Authorization: Bearer <token>。
    - 实体识别无结果统一返回空数组 []，禁止返回 null。
    - 分页统一参数：page（从1开始）、pageSize。
    - 错误响应统一结构见 ErrorVO（code/msg/data，非 200 时 data 为 null）。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: 未登录或 token 缺失/过期（code=401）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVO'
...
    ServerError:
      description: 服务异常（code=500）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVO'
  schemas:
...
    ErrorVO:
      type: object
      description: 统一错误响应（出参；非 200 时 data 为 null）
      properties:
        code:
          type: integer
          description: 错误码：400/401/403/404/500 及业务错误码
        msg:
          type: string
          description: 错误描述
        data:
          type: object
          nullable: true
          example: null
    # ========== 共享数据模型（跨请求/响应复用，非 PO） ==========
    Entity:
      type: object
      description: 实体（含内容与原文片段），共享模型
      properties:
        content:
          type: string
          description: 标准表述
        sourceText:
          type: string
          description: 原文片段
        confidence:
          type: number
          format: double
          description: 置信度（0~1）；模型实体提供，规则兜底可省
        source:
          type: string
          description: 来源：model（NER 模型）/ rule（规则兜底）
          enum:
            - model
            - rule
        normLevel:
          type: integer
          description: 归一命中层级：1=精确 / 2=包含 / 3=模糊；未归一或未命中词典时不返回
        normSource:
          type: string
          description: 归一命中的术语来源（词典名）；未命中不返回
        normCode:
          type: string
          description: 国标代码；词典未收录时不返回
    Herb:
      type: object
      description: 中药实体（含剂量），共享模型
      properties:
        name:
          type: string
          description: 中药名称
        dosage:
          type: string
          description: 剂量，如 15g
        sourceText:
          type: string
          description: 原文片段
        normLevel:
          type: integer
          description: 归一命中层级：1=精确 / 2=包含 / 3=模糊；未归一或未命中词典时不返回
        normSource:
          type: string
          description: 归一命中的术语来源（词典名）；未命中不返回
        normCode:
          type: string
          description: 国标代码；词典未收录时不返回
...
    NlpExtractDTO:
      type: object
      description: NLP 抽取请求（入参）
      required:
        - text
      properties:
        text:
          type: string
          description: 待抽取的病历文本
    NlpExtractVO:
      type: object
      description: NLP 抽取响应（出参，结构同 StructuredData）
      properties:
        diseases:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        symptoms:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        tongueList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        pulseList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        patternList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        causeList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        treatmentList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        formulaList:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        herbs:
          type: array
          items:
            $ref: '#/components/schemas/Herb'
        modelAvailable:
          type: boolean
          description: NLP 服务/模型是否可用；false 表示为降级结果（可能仅规则兜底或为空）
    # ========== LLM 运行时配置 DTO / VO（UX-68） ==========
...
paths:
...
  /api/nlp/extract:
    post:
      tags:
        - NLP抽取
      summary: NLP实体抽取
      description: 【权限：登录即可】NLP实体抽取统一入口，一次返回完整结构化数据（症状/舌脉/证候病因治法/方剂中药）。Java 端转发至 Python FastAPI 服务（:8001），服务/模型不可用时降级（modelAvailable=false）。返回前对实体做术语归一（disease/symptom/pattern/formula/herb 五类走词典，舌/脉/病因/治法四类无独立词典仅保留原文）：content 为标准术语、sourceText 为归一前原文，并附 normLevel/normSource/normCode；未命中词典时 content 保持原文且不返回归一字段
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NlpExtractDTO'
            example:
              text: 患者发热咽痛两天，舌红苔黄，脉浮数。证属风热犯肺，治以辛凉解表，方用银翘散，金银花15g、连翘10g。
      responses:
        '200':
          description: 抽取结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/NlpExtractVO'
        '400':
          description: 待抽取文本为空 / 文本无法解析（乱码或非病历文本）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...